+++ demo/src/main/java/DEMOO/demo/ProductosRepository.java
@@ -3,6 +3,6 @@ package DEMOO.demo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface ProductosRepository extends JpaRepository<Productos, Long> {
+public interface ProductosRepository extends JpaRepository<Productos, Integer> {
 
 }
